Don't try again failed sources
Once an avatar source has failed, it will not be tried again until
a new instance of `AvatarService` is created (typically on page reload).
This avoids repeating the same errors again and again when navigating
between pages in a SPA where a user does not have an available profile
picture (eg: gravatar return 404).

The cache is on a per service and per ID basis. So it is possible to have
some users with working avatar from gravatar, while other users will
try once, fail, and then always fallback on initial avatar instead of gravatar.