Natively lazy load images when scrolling (#102)
This enable native lazy loading of images when user scroll them into view,
as described in https://developer.mozilla.org/en-US/docs/Web/Performance/Lazy_loading

It is especially useful when we have a long list of users and their avatar.
Without this it would start anything between 10-100 http connections to
fetch images that are not visible on screen. With this patch we improve
performance of our app with very little effort and no downsides.